Backend/Microservices/discussion_service/app/main.py
import logging
from fastapi import FastAPI
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
from dotenv import load_dotenv
import os

# Load environment variables
if os.getenv("DOCKER_ENV") == "docker":
    load_dotenv(".env")
else:
    load_dotenv(".env.local")

DATABASE_URL = os.getenv("DATABASE_URL")
print("DATABASE_URL:", DATABASE_URL)  # For debugging

# Configure logging
logging.basicConfig(level=logging.DEBUG)
logger = logging.getLogger(__name__)

engine = create_engine(DATABASE_URL)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
Base = declarative_base()

# Include the router with the correct prefix
from app.routers import discussions, users

app = FastAPI()

app.include_router(discussions.router, prefix="/api/v1", tags=["discussions"])
app.include_router(users.router, prefix="/api/v1", tags=["users"])

@app.get("/")
def read_root():
    return {"message": "Welcome to the Discussion Service API"}

