Backend/Microservices/discussion_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
alembic
httpx
tenacity
os